+++ source/android-library/app/src/main/java/com/google/unity/ads/RewardBasedVideo.java
@@ -99,6 +99,11 @@ public class RewardBasedVideo {
                     public void onRewardedVideoAdLeftApplication() {
                         adListener.onAdLeftApplication();
                     }
+
+                    @Override
+                    public void onRewardedVideoCompleted() {
+                        adListener.onAdCompleted();
+                    }
                 });
             }
         });