+++ source/plugin/Assets/GoogleMobileAds/Api/AdLoader.cs
@@ -104,6 +104,7 @@ namespace GoogleMobileAds.Api
                 get; private set;
             }
 
+
             public Builder ForCustomNativeAd(string templateId)
             {
                 this.TemplateIds.Add(templateId);