+++ source/plugin/Assets/GoogleMobileAds/Common/IRewardBasedVideoAdClient.cs
@@ -34,6 +34,8 @@ namespace GoogleMobileAds.Common
         event EventHandler<EventArgs> OnAdClosed;
         // Ad event fired when the reward based video ad is leaving the application.
         event EventHandler<EventArgs> OnAdLeavingApplication;
+        // Ad event fired when the reward based video ad completes playing.
+        event EventHandler<EventArgs> OnAdCompleted;
 
         // Creates a reward based video ad and adds it to the view hierarchy.
         void CreateRewardBasedVideoAd();