+++ source/plugin/Assets/GoogleMobileAds/Platforms/Android/RewardBasedVideoAdClient.cs
@@ -33,6 +33,7 @@ namespace GoogleMobileAds.Android
         public event EventHandler<EventArgs> OnAdClosed = delegate { };
         public event EventHandler<Reward> OnAdRewarded = delegate { };
         public event EventHandler<EventArgs> OnAdLeavingApplication = delegate { };
+        public event EventHandler<EventArgs> OnAdCompleted = delegate { };
 
         public RewardBasedVideoAdClient()
             : base(Utils.UnityRewardBasedVideoAdListenerClassName)
@@ -148,6 +149,14 @@ namespace GoogleMobileAds.Android
             if (this.OnAdLeavingApplication != null)
             {
                 this.OnAdLeavingApplication(this, EventArgs.Empty);
+            }
+        }
+
+        void onAdCompleted()
+        {
+            if (this.OnAdCompleted != null)
+            {
+                this.OnAdCompleted(this, EventArgs.Empty);
             }
         }