+++ source/plugin/Assets/GoogleMobileAds/Platforms/iOS/Externs.cs
@@ -203,7 +203,8 @@ namespace GoogleMobileAds.iOS
             RewardBasedVideoAdClient.GADURewardBasedVideoAdDidCloseCallback didCloseCallback,
             RewardBasedVideoAdClient.GADURewardBasedVideoAdDidRewardCallback didRewardcallback,
             RewardBasedVideoAdClient.GADURewardBasedVideoAdWillLeaveApplicationCallback
-                    willLeaveCallback);
+                    willLeaveCallback,
+            RewardBasedVideoAdClient.GADURewardBasedVideoAdDidCompleteCallback didCompleteCallback);
 
         [DllImport("__Internal")]
         internal static extern string GADUMediationAdapterClassNameForRewardedVideo(IntPtr rewardedVideo);
@@ -214,7 +215,10 @@ namespace GoogleMobileAds.iOS
 
         [DllImport("__Internal")]
         internal static extern IntPtr GADUCreateAdLoader(
-            IntPtr adLoader, string adUnitId, string[] templateIds, int templateIdsCount,
+            IntPtr adLoader,
+            string adUnitId,
+            string[] templateIds,
+            int templateIdsCount,
             ref NativeAdTypes types);
 
         [DllImport("__Internal")]
@@ -267,47 +271,6 @@ namespace GoogleMobileAds.iOS
             IntPtr nativeCustomTemplateAd,
             CustomNativeTemplateClient.GADUNativeCustomTemplateDidReceiveClick
                     adClickedCallback);
-
-        #endregion
-
-        #region Native Express externs
-
-        [DllImport("__Internal")]
-        internal static extern IntPtr GADUCreateNativeExpressAdView(
-            IntPtr nativeExpressClient, string adUnitId, int width, int height, int positionAtTop);
-
-        [DllImport("__Internal")]
-        internal static extern IntPtr GADUCreateNativeExpressAdViewWithCustomPosition(
-            IntPtr bannerClient,
-            string adUnitId,
-            int width,
-            int height,
-            int x,
-            int y);
-
-        [DllImport("__Internal")]
-        internal static extern void GADUSetNativeExpressAdCallbacks(
-            IntPtr bannerView,
-            NativeExpressAdClient.GADUNativeExpressAdViewDidReceiveAdCallback adReceivedCallback,
-            NativeExpressAdClient.GADUNativeExpressAdViewDidFailToReceiveAdWithErrorCallback adFailedCallback,
-            NativeExpressAdClient.GADUNativeExpressAdViewWillPresentScreenCallback willPresentCallback,
-            NativeExpressAdClient.GADUNativeExpressAdViewDidDismissScreenCallback didDismissCallback,
-            NativeExpressAdClient.GADUNativeExpressAdViewWillLeaveApplicationCallback willLeaveCallback);
-
-        [DllImport("__Internal")]
-        internal static extern void GADUHideNativeExpressAdView(IntPtr nativeExpresAdView);
-
-        [DllImport("__Internal")]
-        internal static extern void GADUShowNativeExpressAdView(IntPtr nativeExpresAdView);
-
-        [DllImport("__Internal")]
-        internal static extern void GADURemoveNativeExpressAdView(IntPtr nativeExpresAdView);
-
-        [DllImport("__Internal")]
-        internal static extern void GADURequestNativeExpressAd(IntPtr nativeExpresAdView, IntPtr request);
-
-        [DllImport("__Internal")]
-        internal static extern string GADUMediationAdapterClassNameForNativeExpressAdView(IntPtr nativeExpressAdView);
 
         #endregion