+++ source/plugin/Assets/Plugins/iOS/GADUAdNetworkExtras.h
@@ -2,9 +2,9 @@
 
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@protocol GADUAdNetworkExtras<NSObject>
+@protocol GADUAdNetworkExtras <NSObject>
 
 - (id<GADAdNetworkExtras>)adNetworkExtrasWithDictionary:
-        (NSDictionary<NSString *, NSString *> *)extras;
+    (NSDictionary<NSString *, NSString *> *)extras;
 
 @end