+++ source/plugin/Assets/Plugins/iOS/GADUInterface.m
@@ -6,7 +6,6 @@
 #import "GADUNativeCustomTemplateAd.h"
 #import "GADUPluginUtil.h"
 #import "GADUAdNetworkExtras.h"
-#import "GADUNativeExpressAd.h"
 #import "GADUObjectCache.h"
 #import "GADURequest.h"
 #import "GADURewardBasedVideoAd.h"
@@ -156,8 +155,9 @@ GADUTypeRewardBasedVideoAdRef GADUCreateRewardBasedVideoAd(
 
 /// Creates a GADUAdLoader and returns its reference.
 GADUTypeAdLoaderRef GADUCreateAdLoader(GADUTypeAdLoaderClientRef *adLoaderClient,
-                                       const char *adUnitID, const char **templateIDs,
-                                       NSInteger templateIDLength, struct AdTypes *types) {
+                                       const char *adUnitID,
+                                       const char **templateIDs, NSInteger templateIDLength,
+                                       struct AdTypes *types) {
   NSMutableArray *templateIDsArray = [[NSMutableArray alloc] init];
   for (int i = 0; i < templateIDLength; i++) {
     [templateIDsArray addObject:GADUStringFromUTF8String(templateIDs[i])];
@@ -166,12 +166,14 @@ GADUTypeAdLoaderRef GADUCreateAdLoader(GADUTypeAdLoaderClientRef *adLoaderClient
   if (types->CustomTemplateAd) {
     [adTypesArray addObject:kGADAdLoaderAdTypeNativeCustomTemplate];
   }
+  NSArray *options = nil;
 
   GADUAdLoader *adLoader =
       [[GADUAdLoader alloc] initWithAdLoaderClientReference:adLoaderClient
                                                    adUnitID:GADUStringFromUTF8String(adUnitID)
                                                 templateIDs:templateIDsArray
-                                                    adTypes:adTypesArray];
+                                                    adTypes:adTypesArray
+                                                    options:options];
   GADUObjectCache *cache = [GADUObjectCache sharedInstance];
   [cache.references setObject:adLoader forKey:[adLoader gadu_referenceKey]];
   return (__bridge GADUTypeAdLoaderRef)adLoader;
@@ -216,7 +218,8 @@ void GADUSetRewardBasedVideoAdCallbacks(
     GADURewardBasedVideoAdDidStartPlayingCallback didStartCallback,
     GADURewardBasedVideoAdDidCloseCallback didCloseCallback,
     GADURewardBasedVideoAdDidRewardCallback didRewardCallback,
-    GADURewardBasedVideoAdWillLeaveApplicationCallback willLeaveCallback) {
+    GADURewardBasedVideoAdWillLeaveApplicationCallback willLeaveCallback,
+    GADURewardBasedVideoAdDidCompleteCallback didCompleteCallback) {
   GADURewardBasedVideoAd *internalRewardBasedVideoAd =
       (__bridge GADURewardBasedVideoAd *)rewardBasedVideoAd;
   internalRewardBasedVideoAd.adReceivedCallback = adReceivedCallback;
@@ -226,56 +229,7 @@ void GADUSetRewardBasedVideoAdCallbacks(
   internalRewardBasedVideoAd.didCloseCallback = didCloseCallback;
   internalRewardBasedVideoAd.didRewardCallback = didRewardCallback;
   internalRewardBasedVideoAd.willLeaveCallback = willLeaveCallback;
-}
-
-/// Sets the native express callback methods to be invoked during native express ad events.
-void GADUSetNativeExpressAdCallbacks(
-    GADUTypeNativeExpressAdRef nativeExpressAd,
-    GADUNativeExpressAdViewDidReceiveAdCallback adReceivedCallback,
-    GADUNativeExpressAdViewDidFailToReceiveAdWithErrorCallback adFailedCallback,
-    GADUNativeExpressAdViewWillPresentScreenCallback willPresentScreenCallback,
-    GADUNativeExpressAdViewDidDismissScreenCallback didDismissScreenCallback,
-    GADUNativeExpressAdViewWillLeaveApplicationCallback willLeaveAppCallback) {
-  GADUNativeExpressAd *internalNativeExpressAd = (__bridge GADUNativeExpressAd *)nativeExpressAd;
-  internalNativeExpressAd.adReceivedCallback = adReceivedCallback;
-  internalNativeExpressAd.adFailedCallback = adFailedCallback;
-  internalNativeExpressAd.willPresentScreenCallback = willPresentScreenCallback;
-  internalNativeExpressAd.didDismissScreenCallback = didDismissScreenCallback;
-  internalNativeExpressAd.willLeaveAppCallback = willLeaveAppCallback;
-}
-
-/// Creates a GADNativeExpressAdView with the specified width, height, and position. Returns
-/// a reference to the GADUNativeExpressAdView.
-GADUTypeNativeExpressAdRef GADUCreateNativeExpressAdView(
-    GADUTypeNativeExpressAdClientRef *nativeExpressAdClient, const char *adUnitID, NSInteger width,
-    NSInteger height, GADAdPosition adPosition) {
-  GADUNativeExpressAd *nativeExpressAd = [[GADUNativeExpressAd alloc]
-      initWithNativeExpressAdClientReference:nativeExpressAdClient
-                                    adUnitID:GADUStringFromUTF8String(adUnitID)
-                                       width:width
-                                      height:height
-                                  adPosition:adPosition];
-
-  GADUObjectCache *cache = [GADUObjectCache sharedInstance];
-  [cache.references setObject:nativeExpressAd forKey:[nativeExpressAd gadu_referenceKey]];
-  return (__bridge GADUTypeNativeExpressAdRef)nativeExpressAd;
-}
-
-/// Creates a GADNativeExpressAdView with the specified width, height, and custom position. Returns
-/// a reference to the GADUNativeExpressAdView.
-GADUTypeNativeExpressAdRef GADUCreateNativeExpressAdViewWithCustomPosition(
-    GADUTypeNativeExpressAdClientRef *nativeExpressAdClient, const char *adUnitID, NSInteger width,
-    NSInteger height, NSInteger x, NSInteger y) {
-  CGPoint adPosition = CGPointMake(x, y);
-  GADUNativeExpressAd *nativeExpressAd = [[GADUNativeExpressAd alloc]
-      initWithNativeExpressAdClientReference:nativeExpressAdClient
-                                    adUnitID:GADUStringFromUTF8String(adUnitID)
-                                       width:width
-                                      height:height
-                            customAdPosition:adPosition];
-  GADUObjectCache *cache = [GADUObjectCache sharedInstance];
-  [cache.references setObject:nativeExpressAd forKey:[nativeExpressAd gadu_referenceKey]];
-  return (__bridge GADUTypeNativeExpressAdRef)nativeExpressAd;
+  internalRewardBasedVideoAd.didCompleteCallback = didCompleteCallback;
 }
 
 /// Sets the banner callback methods to be invoked during native ad events.
@@ -314,24 +268,6 @@ float GADUGetBannerViewHeightInPixels(GADUTypeBannerRef banner) {
 float GADUGetBannerViewWidthInPixels(GADUTypeBannerRef banner) {
   GADUBanner *internalBanner = (__bridge GADUBanner *)banner;
   return internalBanner.widthInPixels;
-}
-
-/// Hides the GADNativeExpressAdView.
-void GADUHideNativeExpressAdView(GADUTypeNativeExpressAdRef nativeExpressAd) {
-  GADUNativeExpressAd *internalNativeExpressAd = (__bridge GADUNativeExpressAd *)nativeExpressAd;
-  [internalNativeExpressAd hideNativeExpressAdView];
-}
-
-/// Shows the GADNativeExpressAdView.
-void GADUShowNativeExpressAdView(GADUTypeBannerRef nativeExpressAd) {
-  GADUNativeExpressAd *internalNativeExpressAd = (__bridge GADUNativeExpressAd *)nativeExpressAd;
-  [internalNativeExpressAd showNativeExpressAdView];
-}
-
-/// Removes the GADNativeExpressAdView from the view hierarchy.
-void GADURemoveNativeExpressAdView(GADUTypeNativeExpressAdRef nativeExpressAd) {
-  GADUNativeExpressAd *internalNativeExpressAd = (__bridge GADUNativeExpressAd *)nativeExpressAd;
-  [internalNativeExpressAd removeNativeExpressAdView];
 }
 
 /// Returns YES if the GADInterstitial is ready to be shown.
@@ -493,14 +429,6 @@ void GADURequestRewardBasedVideoAd(GADUTypeRewardBasedVideoAdRef rewardBasedVide
                              withAdUnitID:GADUStringFromUTF8String(adUnitID)];
 }
 
-/// Makes a native express ad request.
-void GADURequestNativeExpressAd(GADUTypeNativeExpressAdRef nativeExpressAd,
-                                GADUTypeRequestRef request) {
-  GADUNativeExpressAd *internalNativeExpressAd = (__bridge GADUNativeExpressAd *)nativeExpressAd;
-  GADURequest *internalRequest = (__bridge GADURequest *)request;
-  [internalNativeExpressAd loadRequest:[internalRequest request]];
-}
-
 /// Makes a native ad request.
 void GADURequestNativeAd(GADUTypeAdLoaderRef adLoader, GADUTypeRequestRef request) {
   GADUAdLoader *internalAdLoader = (__bridge GADUAdLoader *)adLoader;
@@ -600,12 +528,6 @@ void GADURelease(GADUTypeRef ref) {
 const char *GADUMediationAdapterClassNameForBannerView(GADUTypeBannerRef bannerView) {
   GADUBanner *banner = (__bridge GADUBanner *)bannerView;
   return cStringCopy(banner.mediationAdapterClassName.UTF8String);
-}
-
-const char *GADUMediationAdapterClassNameForNativeExpressAdView(
-    GADUTypeNativeExpressAdRef nativeExpressAdView) {
-  GADUNativeExpressAd *nativeAd = (__bridge GADUNativeExpressAd *)nativeExpressAdView;
-  return cStringCopy(nativeAd.mediationAdapterClassName.UTF8String);
 }
 
 const char *GADUMediationAdapterClassNameForRewardedVideo(