+++ source/plugin/Assets/Plugins/iOS/GADURewardBasedVideoAd.m
@@ -119,4 +119,10 @@
   }
 }
 
+- (void)rewardBasedVideoAdDidCompletePlaying:(GADRewardBasedVideoAd *)rewardBasedVideoAd {
+  if (self.didCompleteCallback) {
+    self.didCompleteCallback(self.rewardBasedVideoAdClient);
+  }
+}
+
 @end