+++ source/plugin/Assets/Plugins/iOS/GADUTypes.h
@@ -32,9 +32,6 @@ typedef const void *GADUTypeAdLoaderClientRef;
 /// Type representing a Unity native custom template ad.
 typedef const void *GADUTypeNativeCustomTemplateAdClientRef;
 
-/// Type representing a Unity native express ad client.
-typedef const void *GADUTypeNativeExpressAdClientRef;
-
 /// Type representing a GADUBanner.
 typedef const void *GADUTypeBannerRef;
 
@@ -49,9 +46,6 @@ typedef const void *GADUTypeAdLoaderRef;
 
 /// Type representing a GADUNativeCustomTemplateAd.
 typedef const void *GADUTypeNativeCustomTemplateAdRef;
-
-/// Type representing a GADUNativeExpressAd.
-typedef const void *GADUTypeNativeExpressAdRef;
 
 /// Type representing a GADURequest.
 typedef const void *GADUTypeRequestRef;
@@ -135,31 +129,9 @@ typedef void (*GADURewardBasedVideoAdDidRewardCallback)(
 typedef void (*GADURewardBasedVideoAdWillLeaveApplicationCallback)(
     GADUTypeRewardBasedVideoAdClientRef *rewardBasedVideoClient);
 
-/// Callback for when a native express ad request was successfully loaded.
-typedef void (*GADUNativeExpressAdViewDidReceiveAdCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient);
-
-/// Callback for when a native express ad request failed.
-typedef void (*GADUNativeExpressAdViewDidFailToReceiveAdWithErrorCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient, const char *error);
-
-/// Callback for when a full screen view is about to be presented as a result of a native express
-/// click.
-typedef void (*GADUNativeExpressAdViewWillPresentScreenCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient);
-
-/// Callback for when a full screen view is about to be dismissed.
-typedef void (*GADUNativeExpressAdViewWillDismissScreenCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient);
-
-/// Callback for when a full screen view has just been dismissed.
-typedef void (*GADUNativeExpressAdViewDidDismissScreenCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient);
-
-/// Callback for when an application will background or terminate as a result of a native express
-/// click.
-typedef void (*GADUNativeExpressAdViewWillLeaveApplicationCallback)(
-    GADUTypeNativeExpressAdClientRef *nativeExpressClient);
+/// Callback for when a reward based video ad completes playing.
+typedef void (*GADURewardBasedVideoAdDidCompleteCallback)(
+    GADUTypeRewardBasedVideoAdClientRef *rewardBasedVideoClient);
 
 /// Callback for when a native custom template ad request was successfully loaded.
 typedef void (*GADUAdLoaderDidReceiveNativeCustomTemplateAdCallback)(