+++ mediation/AdColony/CHANGELOG.md
@@ -1,7 +1,10 @@
 # AdColony Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
-## 1.0.0
+## 1.0.1
+- Supports AdColony Android SDK version 3.3.0-unity.
+- Supports AdColony iOS SDK version 3.3.0.
 
+## 1.0.0
 - First release!
-- Supports Android adapter version 3.3.0.0.
-- Supports iOS adapter version 3.3.0.0.
+- Supports AdColony Android SDK version 3.3.0.
+- Supports AdColony iOS SDK version 3.3.0.