+++ mediation/AppLovin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # AppLovin Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 1.2.1
+- Supports AppLovin Android SDK version 7.8.6.
+- Supports AppLovin iOS SDK version 4.8.3.
+
+## 1.2.0
+- Supports AppLovin Android SDK version 7.8.5.
+- Supports AppLovin iOS SDK version 4.8.3.
+
 ## 1.1.0
 - Supports AppLovin Android SDK version 7.7.0.
 - Supports AppLovin iOS SDK version 4.7.0.