+++ mediation/MyTarget/CHANGELOG.md
@@ -1,7 +1,14 @@
 # myTarget Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
-## 1.0.0
+## 2.0.0
+- Supports myTarget Android SDK version 5.0.0.
+- Supports myTarget iOS SDK version 4.7.7.
 
+## 1.1.0
+- Supports myTarget Android SDK version 4.7.2.
+- Supports myTarget iOS SDK version 4.7.6.
+
+## 1.0.0
 - First release!
-- Supports Android adapter version 4.6.27.0.
-- Supports iOS adapter version 4.6.24.0.
+- Supports myTarget Android SDK version 4.6.27.
+- Supports myTarget iOS SDK version 4.6.24.