+++ mediation/UnityAds/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unity Ads Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
-## 1.0.0
+## 1.1.0
+- Supports Unity Ads Android SDK version 2.2.0.
+- Supports Unity Ads iOS SDK version 2.2.0.
 
+## 1.0.0
 - First release!
-- Supports Android adapter version 2.1.2.0.
-- Supports iOS adapter version 2.1.2.0.
+- Supports Unity Ads Android SDK version 2.1.2.
+- Supports Unity Ads iOS SDK version 2.1.2.